refactor(judicial-system): update broken imports
