+++ apps/judicial-system/web/pages/domur/rannsoknarheimild/fyrirtaka/[id].ts
@@ -1,3 +1,3 @@
-import HearingArrangements from '@island.is/judicial-system-web/src/routes/Court/InvestigationRequest/HearingArrangements/HearingArrangements'
+import HearingArrangements from '@island.is/judicial-system-web/src/routes/Court/InvestigationCase/HearingArrangements/HearingArrangements'
 
 export default HearingArrangements
