+++ apps/judicial-system/web/pages/domur/rannsoknarheimild/stadfesta/[id].ts
@@ -1,3 +1,3 @@
-import Confirmation from '@island.is/judicial-system-web/src/routes/Court/InvestigationRequest/Confirmation/Confirmation'
+import Confirmation from '@island.is/judicial-system-web/src/routes/Court/InvestigationCase/Confirmation/Confirmation'
 
 export default Confirmation
