+++ apps/judicial-system/web/pages/domur/rannsoknarheimild/thingbok/[id].ts
@@ -1,3 +1,3 @@
-import CourtRecord from '@island.is/judicial-system-web/src/routes/Court/InvestigationRequest/CourtRecord/CourtRecord'
+import CourtRecord from '@island.is/judicial-system-web/src/routes/Court/InvestigationCase/CourtRecord/CourtRecord'
 
 export default CourtRecord
