+++ apps/judicial-system/web/pages/domur/rannsoknarheimild/urskurdur/[id].ts
@@ -1,3 +1,3 @@
-import Ruling from '@island.is/judicial-system-web/src/routes/Court/InvestigationRequest/Ruling/Ruling'
+import Ruling from '@island.is/judicial-system-web/src/routes/Court/InvestigationCase/Ruling/Ruling'
 
 export default Ruling
