+++ apps/judicial-system/web/pages/domur/rannsoknarheimild/yfirlit/[id].ts
@@ -1,3 +1,3 @@
-import Overview from '@island.is/judicial-system-web/src/routes/Court/InvestigationRequest/Overview/Overview'
+import Overview from '@island.is/judicial-system-web/src/routes/Court/InvestigationCase/Overview/Overview'
 
 export default Overview
