+++ apps/judicial-system/web/pages/krafa/rannsoknarheimild/domkrofur-og-lagaakvaedi/[id].ts
@@ -1,2 +1,2 @@
-import PoliceDemands from '@island.is/judicial-system-web/src/routes//Prosecutor/InvestigationRequest/PoliceDemands/PoliceDemands'
+import PoliceDemands from '@island.is/judicial-system-web/src/routes//Prosecutor/InvestigationCase/PoliceDemands/PoliceDemands'
 export default PoliceDemands
