+++ apps/judicial-system/web/pages/krafa/rannsoknarheimild/fyrirtaka/[id].ts
@@ -1,2 +1,2 @@
-import HearingArrangements from '@island.is/judicial-system-web/src/routes//Prosecutor/InvestigationRequest/HearingArrangements/HearingArrangements'
+import HearingArrangements from '@island.is/judicial-system-web/src/routes//Prosecutor/InvestigationCase/HearingArrangements/HearingArrangements'
 export default HearingArrangements
