+++ apps/judicial-system/web/pages/krafa/rannsoknarheimild/greinargerd/[id].ts
@@ -1,2 +1,2 @@
-import PoliceReport from '@island.is/judicial-system-web/src/routes//Prosecutor/InvestigationRequest/PoliceReport/PoliceReport'
+import PoliceReport from '@island.is/judicial-system-web/src/routes//Prosecutor/InvestigationCase/PoliceReport/PoliceReport'
 export default PoliceReport
