+++ apps/judicial-system/web/pages/krafa/rannsoknarheimild/index.ts
@@ -1,3 +1,3 @@
-import Defendant from '@island.is/judicial-system-web/src/routes//Prosecutor/InvestigationRequest/Defendant/Defendant'
+import Defendant from '@island.is/judicial-system-web/src/routes//Prosecutor/InvestigationCase/Defendant/Defendant'
 
 export default Defendant
