+++ apps/judicial-system/web/pages/krafa/rannsoknarheimild/rannsoknargogn/[id].ts
@@ -1,2 +1,2 @@
-import CaseFiles from '@island.is/judicial-system-web/src/routes//Prosecutor/InvestigationRequest/CaseFiles/CaseFiles'
+import CaseFiles from '@island.is/judicial-system-web/src/routes//Prosecutor/InvestigationCase/CaseFiles/CaseFiles'
 export default CaseFiles
