+++ apps/judicial-system/web/pages/krafa/rannsoknarheimild/stadfesta/[id].ts
@@ -1,2 +1,2 @@
-import { Overview } from '@island.is/judicial-system-web/src/routes//Prosecutor/InvestigationRequest/Overview/Overview'
+import { Overview } from '@island.is/judicial-system-web/src/routes//Prosecutor/InvestigationCase/Overview/Overview'
 export default Overview
