+++ apps/judicial-system/web/pages/krafa/rannsoknarheimild/varnaradili/[id].ts
@@ -1,2 +1,2 @@
-import Defendant from '@island.is/judicial-system-web/src/routes//Prosecutor/InvestigationRequest/Defendant/Defendant'
+import Defendant from '@island.is/judicial-system-web/src/routes//Prosecutor/InvestigationCase/Defendant/Defendant'
 export default Defendant
