+++ apps/judicial-system/web/src/routes/Shared/Cases/TableSkeleton.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 
 import { SkeletonLoader, Text, Box } from '@island.is/island-ui/core'
 
-import * as styles from './Requests.css'
+import * as styles from './Cases.css'
 
 const TableSkeleton = () => {
   return (
